requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pyyaml
python-dotenv
streamlit
plotly
pandas
